requirements.txt
Django==5.0.6
gunicorn==22.0.0
djangorestframework==3.15.2
django-rest-swagger==2.2.0